Add badges and refactor app structure
